refactor(lms): 기간에 대해서 dateRange 객체를 생성하여 처리
- Session 에 존재하던 시작일, 종료일에 대한 유효성 검증 책임을 DateRange 에 위임
- 해당 테스트코드 이동
- 문서 수정
