+++ STEP2-README.md
@@ -26,10 +26,13 @@
 - [x] 기수가 1보다 작은 수가 지정 되었다면 예외가 발생한다.
 - [x] 기수와 Session 을 포함한다면 Course 는 생성된다. 
 ### Session
-- [x] 시작일과 종료일 중 어느하나라도 설정되지 않은경우 예외가 발생한다.
-- [x] 종료일은 시작일 이전으로 설정이 불가능하다.
+- [ ] 모집시작일과 모집종료일은 필수이다.
+- [ ] 모집종료일은 모집시작일 이전으로 설정이 불가능하다.
 - [ ] Session 이 '모집중' 상태가 아니면 신청 시 예외가 발생한다.
 - [ ] 유료강의인 경우 수강생이 결제한 금액과 수강료가 일치하면 신청된다.
+### DateRange
+- [x] 시작일과 종료일은 필수이다.
+- [x] 종료일은 시작일 이전으로 설정이 불가능하다.
 ### SessionCoverImage
 - [ ] 1MB 를 초과하는 파일이 전달된 경우 예외가 발생한다.
 - [ ] 지정된 확장자 외의 확장자가 전달된 경우 예외가 발생한다.
