+++ src/test/java/nextstep/courses/domain/CourseTest.java
@@ -1,7 +1,5 @@
-package nextstep.courses;
+package nextstep.courses.domain;
 
-import nextstep.courses.domain.CourseV2;
-import nextstep.courses.domain.Session;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
@@ -30,11 +28,10 @@ public class CourseTest {
     @Test
     @DisplayName("기수와 Session 을 포함한다면 Course 는 생성된다.")
     void shouldCreateCourseWhenCohortAndSessionAreProvided() {
-        final LocalDate startDate = LocalDate.of(2024, 5, 30);
-        final LocalDate endDate = LocalDate.of(2024, 10, 29);
+        final DateRange dateRange = new DateRange(LocalDate.of(2024, 5, 30), LocalDate.of(2024, 10, 29));
 
         final CourseV2 courseV2 = new CourseV2("TDD, 클린 코드 with Java", 40, List.of(
-            new Session(startDate, endDate)
+            new Session(dateRange)
         ));
 
         assertThat(courseV2).isNotNull();
