+++ src/test/java/nextstep/courses/domain/DateRangeTest.java
@@ -1,6 +1,5 @@
-package nextstep.courses;
+package nextstep.courses.domain;
 
-import nextstep.courses.domain.Session;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
@@ -8,12 +7,12 @@ import java.time.LocalDate;
 
 import static org.assertj.core.api.Assertions.assertThatIllegalArgumentException;
 
-public class SessionTest {
+public class DateRangeTest {
     @Test
     @DisplayName("시작일과 종료일 중 어느하나라도 설정되지 않은 경우 예외가 발생한다.")
     void shouldThrowExceptionWhenStartDateOrEndDateIsMissing() {
         assertThatIllegalArgumentException()
-            .isThrownBy(() -> new Session(LocalDate.now(), null));
+            .isThrownBy(() -> new DateRange(LocalDate.now(), null));
     }
 
     @Test
@@ -23,6 +22,6 @@ public class SessionTest {
         final LocalDate endDate = LocalDate.of(2024, 1, 1);
 
         assertThatIllegalArgumentException()
-            .isThrownBy(() -> new Session(startDate, endDate));
+            .isThrownBy(() -> new DateRange(startDate, endDate));
     }
 }
