+++ src/test/java/nextstep/courses/domain/SessionTest.java
@@ -1,0 +1,4 @@
+package nextstep.courses.domain;
+
+public class SessionTest {
+}
